LayoutTests/fast/shadow-dom/slot-display-contents-composed-tree-after-display-toggle-expected.txt

Composed tree from #mutate:
      div
        #text
      #text
      div
        #text
      #text
    #text
A
B
